feature: rename a file
